o several bugfixes in instationary adjoint equations and the data handling between fluid and adjoint fluid   -> the adjoints are (hopefully) bugfree now, waiting to be used for optimization :)
o nightly test cases for instationary topopt adjoint equations

o made several variables constant

o a little less header inclusions in the adapters
o removed some build warnings by minor modifications in several files

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16245 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_lib/drt_utils_parmetis.cpp
@@ -1029,8 +1029,8 @@ void DRT::UTILS::PartUsingParMetis(RCP<DRT::Discretization> dis,
     int npart = numproc;         // number of partitions desired
     int options[4] = { 0,0,15,0 }; // use default metis parameters
     int edgecut = 0;             // output, number of edges cut in partitioning
-    float ubvec = 1.05;
-    vector<float> tpwgts(npart,1.0/(double)npart);
+    float ubvec = (float)1.05;
+    vector<float> tpwgts(npart,static_cast<float>(1.0/(double)npart));
     MPI_Comm mpicomm=(dynamic_cast<const Epetra_MpiComm*>(&(dis->Comm())))->Comm();
 
     ParMETIS_V3_PartKway(&(vtxdist[0]),&(xadj[0]),&(adjncy[0]),
